requirements.txt
selenium
sqlalchemy
python-dotenv